pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["dari"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
envparse = "^0.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
social-auth-app-django = "^5.0.0"
psycopg2-binary = "^2.9.5"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
